Code fonctionnel + Checkstyle

+++ src/main/java/fr/uvsq21504875/CarreDAO.java
@@ -7,7 +7,17 @@ import java.util.List;
 
 public class CarreDAO extends DAO<Carre> {
 
+  /**
+   * Variable de connexion à la base de données.
+   */
   private DBConnection db = new DBConnection();
+
+  /**
+   * Méthode de création d'un Carré dans la base de données.
+   *
+   * @param obj L'objet à créer dans la base de données.
+   * @return Le carré crée
+   */
   @Override
   public Carre create(final Carre obj) {
     db.connect();
@@ -18,7 +28,7 @@ public class CarreDAO extends DAO<Carre> {
       ps.setString(1, obj.getNom());
       ps.setInt(2, obj.getHG().getX());
       ps.setInt(3, obj.getHG().getY());
-      ps.setDouble(4, obj.getCote());
+      ps.setInt(4, obj.getCote());
       int result = ps.executeUpdate();
       assert result == 1;
       ps.close();
@@ -29,6 +39,12 @@ public class CarreDAO extends DAO<Carre> {
     return obj;
   }
 
+  /**
+   * Méthode de recherche d'un Carré.
+   *
+   * @param id Le paramètre d'identification de l'objet.
+   * @return Le Carré trouvé (ou null).
+   */
   @Override
   public Carre find(final String id) {
     db.connect();
@@ -39,7 +55,7 @@ public class CarreDAO extends DAO<Carre> {
       );
       ps.setString(1, id);
       ResultSet result = ps.executeQuery();
-      if(result.first()) {
+      if (result.first()) {
         c1 = new Carre(
             result.getString("nom"),
             result.getInt("HGx"),
@@ -55,16 +71,21 @@ public class CarreDAO extends DAO<Carre> {
     return c1;
   }
 
+  /**
+   * Fonction de récupération du contenu de la Table Carre.
+   *
+   * @return Les Carrés sauvegardés.
+   */
   @Override
   public List<Carre> findAll() {
     db.connect();
     List<Carre> lc = new ArrayList<>();
-    try{
+    try {
       PreparedStatement ps = db.getConn().prepareStatement(
           "SELECT * FROM Carres"
       );
       ResultSet result = ps.executeQuery();
-      while(result.next()){
+      while (result.next()) {
         lc.add(new Carre(
             result.getString("nom"),
             result.getInt("HGx"),
@@ -73,13 +94,19 @@ public class CarreDAO extends DAO<Carre> {
         ));
       }
       ps.close();
-    }catch(Exception e){
+    } catch (Exception e) {
       e.printStackTrace();
     }
     db.disconnect();
     return lc;
   }
 
+  /**
+   * Méthode de mise à jour d'un objet de la Table.
+   *
+   * @param obj L"objet à mettre à jour.
+   * @return L'objet modifié.
+   */
   @Override
   public Carre update(final Carre obj) {
     db.connect();
@@ -101,7 +128,11 @@ public class CarreDAO extends DAO<Carre> {
     db.disconnect();
     return obj;
   }
-
+  /**
+   * Méthode de suppression de la Table.
+   *
+   * @param obj L"objet à supprimer.
+   */
   @Override
   public void delete(final Carre obj) {
     db.connect();
